requirements.txt
pyspark==3.1.1
coloredlogs
Faker
confluent-kafka
kafka-python